Refactor express checkout element interactor
Committed-By-Agent: codex
