+++ paymentsheet/src/main/java/com/stripe/android/checkout/ExpressCheckoutElement.kt
@@ -3,8 +3,6 @@ package com.stripe.android.checkout
 import android.os.Parcelable
 import androidx.annotation.RestrictTo
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import com.stripe.android.checkout.ece.DefaultExpressCheckoutElementInteractor
 import com.stripe.android.checkout.ece.ExpressCheckoutElementContent
 import com.stripe.android.checkout.ece.ExpressCheckoutElementInteractor
 import com.stripe.android.paymentelement.CheckoutSessionPreview
@@ -14,13 +12,11 @@ import kotlinx.parcelize.Parcelize
 @CheckoutSessionPreview
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class ExpressCheckoutElement internal constructor(
-    private val interactorFactory: ExpressCheckoutElementInteractor.Factory =
-        DefaultExpressCheckoutElementInteractor.Factory,
+    private val interactor: ExpressCheckoutElementInteractor,
 ) {
 
     @Composable
     fun Content() {
-        val interactor = remember(interactorFactory) { interactorFactory.create() }
         ExpressCheckoutElementContent(interactor = interactor)
     }
