+++ paymentsheet/src/main/java/com/stripe/android/checkout/ece/ExpressButton.kt
@@ -1,0 +1,6 @@
+package com.stripe.android.checkout.ece
+
+internal sealed interface ExpressButton {
+    data object Link : ExpressButton
+    data object GooglePay : ExpressButton
+}
