+++ paymentsheet/src/test/java/com/stripe/android/checkout/ece/DefaultExpressCheckoutElementInteractorTest.kt
@@ -5,19 +5,16 @@ import org.junit.Test
 
 internal class DefaultExpressCheckoutElementInteractorTest {
     @Test
-    fun `default state contains no wallet buttons`() {
+    fun `default state contains link and google pay buttons`() {
         val interactor = DefaultExpressCheckoutElementInteractor()
 
         assertThat(interactor.state.value).isEqualTo(
-            ExpressCheckoutElementInteractor.State(walletButtons = emptyList())
+            ExpressCheckoutElementInteractor.State(
+                expressButtons = listOf(
+                    ExpressButton.Link,
+                    ExpressButton.GooglePay,
+                )
+            )
         )
-    }
-
-    @Test
-    fun `factory creates default interactor`() {
-        val interactor = DefaultExpressCheckoutElementInteractor.Factory.create()
-
-        assertThat(interactor).isInstanceOf(DefaultExpressCheckoutElementInteractor::class.java)
-        assertThat(interactor.state.value.walletButtons).isEmpty()
     }
 }
